fix(profiling): avoid implicit widening of multiplication result
Use explicit static_cast<size_t> to avoid implicit widening
conversion when multiplying int literals assigned to size_t.

Fixes clang-tidy bugprone-implicit-widening-of-multiplication-result warnings.

+++ ddtrace/internal/datadog/profiling/stack/src/echion/vm.cc
@@ -78,7 +78,7 @@ VmReader*
 VmReader::get_instance()
 {
     if (instance == nullptr) {
-        instance = VmReader::create(1024 * 1024); // A megabyte?
+        instance = VmReader::create(static_cast<size_t>(1024) * 1024); // A megabyte?
         if (!instance) {
             std::cerr << "Failed to initialize VmReader with buffer size " << instance->sz << std::endl;
             return nullptr;
